fix: correct reassign logic for workflow_run in data sync
