+++ server/application-server/src/main/java/de/tum/cit/aet/helios/github/GitHubService.java
@@ -12,14 +12,20 @@ import de.tum.cit.aet.helios.filters.RepositoryContext;
 import de.tum.cit.aet.helios.github.permissions.GitHubPermissionsResponse;
 import de.tum.cit.aet.helios.github.permissions.GitHubRepositoryRoleDto;
 import de.tum.cit.aet.helios.github.permissions.RepoPermissionType;
+import de.tum.cit.aet.helios.workflow.GitHubWorkflowContext;
 import jakarta.transaction.Transactional;
+import java.io.BufferedReader;
+import java.io.FilterInputStream;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.time.OffsetDateTime;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import okhttp3.MediaType;
@@ -597,4 +603,132 @@ public class GitHubService {
       return getRepository(repoNameWithOwner).getReleaseByTagName(tagName);
     }
   }
+
+  /**
+   * Extracts workflow context from the workflow-context artifact.
+   *
+   * @param repositoryId The ID of the repository
+   * @param runId The ID of the workflow run
+   * @return The extracted GitHubWorkflowContext or null if not found or error occurred
+   */
+  public GitHubWorkflowContext extractWorkflowContext(long repositoryId, long runId) {
+    GHArtifact ghArtifact = null;
+
+    // Fetch artifacts to get the triggering workflow run ID
+    try {
+      PagedIterable<GHArtifact> artifacts =
+          getWorkflowRunArtifacts(repositoryId, runId);
+
+      // First artifact with the configured name
+      for (GHArtifact artifact : artifacts) {
+        if (artifact.getName().equals("workflow-context")) {
+          ghArtifact = artifact;
+          break;
+        }
+      }
+    } catch (Exception e) {
+      log.warn("Failed to fetch artifacts for workflow run: {}", e.getMessage());
+      return null;
+    }
+
+    if (ghArtifact == null) {
+      log.warn("No workflow-context artifact found for E2E Tests workflow_run: {}", runId);
+      return null;
+    }
+
+    log.debug("Found artifact {}", ghArtifact.getName());
+
+    try {
+      // Parse the artifact to extract the triggering workflow information
+      return parseWorkflowContextArtifact(ghArtifact);
+    } catch (Exception e) {
+      log.error("Failed to parse workflow context artifact: {}", e.getMessage());
+      return null;
+    }
+  }
+
+  /**
+   * Parses the workflow context artifact to extract the triggering workflow information.
+   */
+  private GitHubWorkflowContext parseWorkflowContextArtifact(GHArtifact artifact)
+      throws IOException {
+    // Download & Parse the artifact
+    return artifact.download(
+        artifactContent -> {
+          if (artifactContent.available() == 0) {
+            throw new RuntimeException("Empty artifact stream!");
+          }
+
+          Long workflowRunId = null;
+          String headBranch = null;
+          String headSha = null;
+
+          try (ZipInputStream zipInput = new ZipInputStream(artifactContent)) {
+            ZipEntry entry;
+
+            while ((entry = zipInput.getNextEntry()) != null) {
+              if (!entry.isDirectory()
+                  && "workflow-context.txt".equalsIgnoreCase(entry.getName())) {
+
+                var nonClosingStream =
+                    new FilterInputStream(zipInput) {
+                      @Override
+                      public void close() throws IOException {
+                        // Do nothing, so the underlying stream stays open.
+                      }
+                    };
+
+                // Read file content
+                try (BufferedReader reader =
+                         new BufferedReader(new InputStreamReader(nonClosingStream))) {
+                  String line;
+                  while ((line = reader.readLine()) != null) {
+                    // Skip empty lines
+                    if (line.trim().isEmpty()) {
+                      continue;
+                    }
+
+                    // Split each line by equals sign
+                    String[] parts = line.split("=", 2);
+                    if (parts.length == 2) {
+                      String key = parts[0].trim();
+                      String value = parts[1].trim();
+
+                      // Extract values
+                      switch (key) {
+                        case "TRIGGERING_WORKFLOW_RUN_ID" -> workflowRunId = Long.parseLong(value);
+                        case "TRIGGERING_WORKFLOW_HEAD_BRANCH" -> headBranch = value;
+                        case "TRIGGERING_WORKFLOW_HEAD_SHA" -> headSha = value;
+                        default -> log.warn("Unknown key in workflow-context.txt: {}", key);
+                      }
+                    }
+                  }
+                }
+              }
+              zipInput.closeEntry();
+            }
+          }
+
+          // Validate that we found all required values
+          if (workflowRunId == null) {
+            throw new RuntimeException("Could not find TRIGGERING_WORKFLOW_RUN_ID in artifact");
+          }
+          if (headBranch == null) {
+            throw new RuntimeException(
+                "Could not find TRIGGERING_WORKFLOW_HEAD_BRANCH in artifact");
+          }
+          if (headSha == null) {
+            throw new RuntimeException("Could not find TRIGGERING_WORKFLOW_HEAD_SHA in artifact");
+          }
+
+          log.info(
+              "Context extracted: workflowRunId: {}, headBranch: {}, headSha: {}",
+              workflowRunId,
+              headBranch,
+              headSha);
+
+          return new GitHubWorkflowContext(workflowRunId, headBranch, headSha);
+        });
+  }
+
 }
